backend/requirements.txt
FastAPI
uvicorn[standard]
pydantic
pydantic-setting